requirements.txt
python-dotenv==1.0.1
stripe==11.2.0
pydantic==2.9.2
